Update OS & ItensOS

+++ biontec-api/src/main/java/com/biontecapi/service/ItensOSService.java
@@ -3,7 +3,7 @@ package com.biontecapi.service;
 import com.biontecapi.dtos.ItensDoServicoDTO;
 import com.biontecapi.model.ItensDoServico;
 
-import javax.transaction.Transactional;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -13,7 +13,7 @@ public interface ItensOSService {
 
     List<ItensDoServico> findAll();
 
-    Optional<ItensDoServico> findById(Integer id);
+    Optional<ItensDoServico> findById(Long id);
 
     List<ItensDoServicoDTO> listarItensOSPorIdProduto(Integer id);
